Implement methods for PlayingFieldWidgetQt::Impl

+++ aswui/src/PlayingFieldPresenter.cpp
@@ -1,4 +1,5 @@
 #include "PlayingFieldPresenter.h"
+#include <gsl/narrow>
 
 namespace aswui {
 
@@ -7,8 +8,8 @@ PlayingFieldPresenter::PlayingFieldPresenter(PlayingFieldView& view)
 }
 
 void PlayingFieldPresenter::set(asw::ConstCellSpan const& cells) {
-    view_.set_row_count(cells.extent(0));
-    view_.set_column_count(cells.extent(1));
+    view_.set_row_count(gsl::narrow_cast<int>(cells.extent(0)));
+    view_.set_column_count(gsl::narrow_cast<int>(cells.extent(1)));
 }
 
 }  // namespace aswui
